moved reducers to its own file
